moved test for checkJavaInstallation to testUtilsCommon.js
