+++ tests/unit_tests/commands/android/testIndex.js
@@ -77,102 +77,6 @@ describe('test showHelp', function() {
   });
 });
 
-describe('test checkJavaInstallation', function() {
-  beforeEach(() => {
-    mockery.enable({useCleanCache: true, warnOnReplace: false, warnOnUnregistered: false});
-
-    mockery.registerMock('./adb', {});
-  });
-
-  afterEach(() => {
-    mockery.deregisterAll();
-    mockery.resetCache();
-    mockery.disable();
-  });
-
-  it('returns true if command executed successfully', () => {
-    const consoleOutput = [];
-    mockery.registerMock(
-      '../../../logger',
-      class {
-        static log(...msgs) {
-          consoleOutput.push(...msgs);
-        }
-      }
-    );
-
-    const colorFn = (arg) => arg;
-    mockery.registerMock('ansi-colors', {
-      green: colorFn,
-      yellow: colorFn,
-      magenta: colorFn,
-      cyan: colorFn,
-      red: colorFn,
-      grey: colorFn
-    });
-
-    const commandsExecuted = [];
-    mockery.registerMock('child_process', {
-      execSync(command) {
-        commandsExecuted.push(command);
-      }
-    });
-
-    const {AndroidSetup} = require('../../../../src/commands/android/index');
-    const {checkJavaInstallation} = require('../../../../src/commands/android/utils/common');
-    const androidSetup = new AndroidSetup();
-    const result = checkJavaInstallation(androidSetup.rootDir);
-
-    assert.strictEqual(result, true);
-    assert.strictEqual(commandsExecuted[0], 'java -version');
-
-    const output = consoleOutput.toString();
-    assert.strictEqual(output, '');
-  });
-
-  it('returns false if command execution failed', () => {
-    const consoleOutput = [];
-    mockery.registerMock(
-      '../../../logger',
-      class {
-        static log(...msgs) {
-          consoleOutput.push(...msgs);
-        }
-      }
-    );
-
-    const colorFn = (arg) => arg;
-    mockery.registerMock('ansi-colors', {
-      green: colorFn,
-      yellow: colorFn,
-      magenta: colorFn,
-      cyan: colorFn,
-      red: colorFn,
-      grey: colorFn
-    });
-
-    const commandsExecuted = [];
-    mockery.registerMock('child_process', {
-      execSync(command) {
-        commandsExecuted.push(command);
-        throw Error();
-      }
-    });
-
-    const {AndroidSetup} = require('../../../../src/commands/android/index');
-    const {checkJavaInstallation} = require('../../../../src/commands/android/utils/common');
-    const androidSetup = new AndroidSetup();
-    const result = checkJavaInstallation(androidSetup.rootDir);
-
-    assert.strictEqual(result, false);
-    assert.strictEqual(commandsExecuted[0], 'java -version');
-
-    const output = consoleOutput.toString();
-    assert.strictEqual(output.includes('Java Development Kit v9 or above is required'), true);
-    assert.strictEqual(output.includes('Make sure Java is installed by running java -version'), true);
-  });
-});
-
 describe('test getJavaHomeFromUser', function() {
   beforeEach(() => {
     mockery.enable({useCleanCache: true, warnOnReplace: false, warnOnUnregistered: false});
